first submit

+++ app/Http/Controllers/TasksController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 
 use App\Task;
+use App\User;
 
 class TasksController extends Controller
 {
@@ -15,7 +16,7 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $tasks = \Auth::user()->tasks();
 
         return view('tasks.index', [
             'tasks' => $tasks,
@@ -45,6 +46,8 @@ class TasksController extends Controller
      
     public function store(Request $request)
     {
+        $id=\Auth::id();
+        
         $this->validate($request, [
             'content'=> 'required|max:255',
             "status"=>"required|max:10",
@@ -53,9 +56,10 @@ class TasksController extends Controller
         $task = new Task;
         $task->content = $request->content;
         $task->status = $request->status;
+        $task->user_id=$id;
         $task->save();
         
-        return redirect("/");
+        return redirect()->route("users.show", ["id"=>$id]);
     }
 
     /**
@@ -68,9 +72,15 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
 
+        if (\Auth::id() === $task->user_id) {
         return view('tasks.show', [
             'task' => $task,
         ]);
+        }
+        
+        else {
+            return redirect("/");
+        }
     }
 
     /**
@@ -83,9 +93,13 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
-        return view("tasks.edit", [
-            "task" => $task
-            ]);
+        if (\Auth::id() === $task->user_id) {
+            return view("tasks.edit", ["task"=>$task]);
+        }
+        
+        else {
+                return redirect("/");
+        }
             
     }
 
@@ -98,6 +112,9 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
+        
+        $user_id=\Auth::id();
+        
         $this->validate($request, [
             'content'=> 'required|max:255',
             "status"=> "required|max:10",
@@ -108,7 +125,7 @@ class TasksController extends Controller
         $task->status = $request->status;
         $task->save();
         
-        return redirect("/"); 
+        return redirect()->route("users.show", ["id"=>$user_id]);
     }
 
     /**
@@ -119,9 +136,17 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $user_id=\Auth::id();
         
-        return redirect("/");
+        $task = Task::find($id);
+        
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect()->route("users.show", ["id"=>$user_id]);
+        }
+        
+        else {
+            return redirect("/");
+        }
     }
 }
